+++ edgy/core/db/fields/foreign_keys.py
@@ -188,12 +188,30 @@ class BaseForeignKeyField(BaseForeignKey):
 
         # If the value is already a target model instance or its proxy.
         if isinstance(value, target | target.proxy_model):
+            maybe_loop = False
+            db_dirty = value._db_dirty
             # Save the related model first to ensure its primary key is available.
-            await value.real_save(force_insert=False, values=None)
+            if db_dirty is None:
+                maybe_loop = True
+            else:
+                # when None _db_dirty is disabled so don't try to save
+                if not value._db_loaded or not value.can_load:
+                    # save when not loaded (could affect relationship)
+                    # required; otherwise save loops can happen.
+                    await value.real_save(force_insert=False, values=None)
+                else:
+                    # save when fields were changed (could affect relationship)
+                    await value.real_save(force_insert=False, values=db_dirty)
             # Clean the value to extract the foreign key column values.
             _value: dict[str, Any] = self.clean(self.name, value, for_query=False)
             # it is an error if it couldn't be parsed.
-            if isinstance(_value.get(self.name), BaseModelType):
+            if isinstance(_value.get(self.name), BaseModelType) or (
+                original_value is not None and not _value
+            ):
+                if maybe_loop:
+                    raise ValueError(
+                        f"Couldn't save `{self.name}`. Please save child first. Is a loop?"
+                    )
                 raise ValueError(f"Couldn't save `{self.name}`.")
             return _value
         # If the value is a dictionary, convert it to a target model instance and
@@ -237,7 +255,11 @@ class BaseForeignKeyField(BaseForeignKey):
         return cast(
             ManyRelationProtocol,
             relation(
-                to=self.owner, to_foreign_key=self.name, embed_parent=self.embed_parent, **kwargs
+                to=self.owner,
+                reverse_name=self.reverse_name,
+                to_foreign_key=self.name,
+                embed_parent=self.embed_parent,
+                **kwargs,
             ),
         )