+++ docs/release-notes.md
@@ -13,6 +13,7 @@
 - Add `SkipOperation` exception for signals.
 - Add `injected_filters` parameter for `pre_delete` to dynamically inject protection rules.
 - Allow lazy references for `ForeignKey` and `ManyToMany`. This is useful for libraries.
+- Add dirty tracking.
 
 ### Changed
 
@@ -28,6 +29,7 @@
 - Refactor `QueryExecutor` so it can update itself.
 - `real_save`, `raw_delete` and `delete` are now keyword only to prevent bad overwrites.
 - Remove the `skip_post_delete_hooks` parameter from `delete` as it is for internal use and shouldn't be exposed.
+- (Internal detail only) bypass the overwritten `__getattr__`, `__getattribute__` and `__setattr__` for speed improvements.
 
 ### Fixed